Upgraded README.md and formatted code
The README.md is now the "Schriftliche Arbeit" I submitted at
www.jugend-forscht.de and I formatted my code with the tool black
